Add SelectionOrderedSet abstraction
This encompasses much of the logic of both Dock and DockGroup. In fact,
it also encompasses much of PaneContainer and Pane. In the future, we
should probably try to rewrite them in terms of this but, given that
there are subtle differences in behavior, we'll leave that for another
day.